changed file name and package for unit tests
